Consolidate endpoint logic to endpoints.js
